Use keymap-introspection instead of accessing keymaps directly
This fixes an "error: 'keymaps' undeclared" error.

It also highlights the fact that we discard the uppermost 8 bits of
the keycode when filling in the openrgb_get_led_info result.  This is
problematic if you e.g. assign QK_BOOTLOADER (0X7C00) to the top-left
key in the default layer, as that will fill in 0 in the result, which
is interpreted as KC_NO (0x0000).  So avoid doing that, I guess...